backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.1
psycopg2-binary>=2.9.9

# LangGraph and LangChain
langgraph>=0.2.0
langgraph-checkpoint-postgres>=2.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-google-genai>=2.0.0

# AI/ML
google-generativeai>=0.8.0
openai>=1.12.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0
pillow>=10.2.0
pandas>=2.2.0
openpyxl>=3.1.2

# WebSocket
websockets>=12.0
